Going from lambda to our own server.
Removing the node_modules folder thanks to the dockerization and self hosting.
Addition of a socket in order to send the data to the dashboard. The socket client is in the server replying to Alexa.
Addition of the socket server (wfp-alexa-socket folder) in order to send the data sent by the client to all the dashboard.
Changing the package to write the handlers of the request because of the self hosting.
Addition of the dockers for the socket server and the backend.
Addition of a gulp file to deploy more easily on the reliefappstest server.